+++ org.eclipse.winery.repository.ui/src/app/wineryRepository.component.ts
@@ -6,21 +6,28 @@
  * and are available at http://www.eclipse.org/legal/epl-v20.html
  * and http://www.apache.org/licenses/LICENSE-2.0
  */
-import { Component, ViewChild, ViewContainerRef } from '@angular/core';
+import { Component, OnInit, ViewChild, ViewContainerRef } from '@angular/core';
 import { WineryNotificationService } from './wineryNotificationModule/wineryNotification.service';
 import { WineryGitLogComponent } from './wineryGitLog/wineryGitLog.component';
+import { ExistService } from './wineryUtils/existService';
+import { backendBaseURL } from './configuration';
 
 @Component({
     selector: 'winery-repository',
     templateUrl: './wineryRepository.html',
+    styleUrls: ['./wineryRepository.component.css'],
+    providers: [
+        ExistService
+    ]
 })
 /*
  * This component represents the root component for the Winery Repository.
  */
-export class WineryRepositoryComponent {
+export class WineryRepositoryComponent implements OnInit {
     // region variables
     name = 'Winery Repository';
-
+    isBackendAvailable = false;
+    loading = true;
     @ViewChild('gitLog') gitLog: WineryGitLogComponent;
 
     // endregion
@@ -30,11 +37,26 @@ export class WineryRepositoryComponent {
         lastOnBottom: true
     };
 
-    constructor(vcr: ViewContainerRef, private notify: WineryNotificationService) {
+    constructor(vcr: ViewContainerRef, private notify: WineryNotificationService, private existService: ExistService) {
         this.notify.init(vcr);
+    }
+
+    ngOnInit() {
+        this.existService.check(backendBaseURL).subscribe(
+            data => {
+                this.isBackendAvailable = true;
+                this.loading = false;
+            },
+            error => this.loading = false
+        );
     }
 
     onClick() {
         this.gitLog.hide();
     }
+
+    refresh() {
+        window.location.reload();
+    }
+
 }