+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This is similar to [Angular's CHANGELOG.md](https://github.com/angular/angular/b
 
 ### Changed
 - Add possibility to work with sub directories in artifact templates
+- Add error notification in case of backend is not available
 - Add edit properties as XML
 - Fix lifecycle interface is now selected after click on "generate lifecycle interface"
 - Add available/not available indicator for LFS in git log component